refactor(웹뷰): 폴더구조 변경
