+++ src/components/Basic/BasicModal/BasicModal.ts
@@ -1,5 +1,5 @@
-import BaseComponent from '../BaseComponent';
-
+import BaseComponent from '../../BaseComponent';
+import './BasicModal.css';
 class BasicModal extends BaseComponent {
   #children: Node;
