+++ src/components/RestaurantList/RestaurantItem/RestaurantItem.ts
@@ -1,4 +1,4 @@
-import BaseComponent from '../BaseComponent';
+import BaseComponent from '../../BaseComponent';
 import koreanIcon from '@/assets/category-korean.png';
 import asianIcon from '@/assets/category-asian.png';
 import japaneseIcon from '@/assets/category-japanese.png';
@@ -7,6 +7,7 @@ import westernIcon from '@/assets/category-western.png';
 import etcIcon from '@/assets/category-etc.png';
 import { Category, IRestaurant } from '@/types/Restaurant';
 
+import './RestaurantItem.css';
 export const Icons: { [key: Category]: string } = {
   한식: koreanIcon,
   아시안: asianIcon,
