+++ src/components/RestaurantList/RestaurantList.ts
@@ -1,8 +1,9 @@
 import RestaurantDBService from '@/domains/services/RestaurantDBService';
 import BaseComponent from '../BaseComponent';
-import RestaurantItem from '../RestaurantItem/RestaurantItem';
+import RestaurantItem from './RestaurantItem/RestaurantItem';
 import { IRestaurant } from '@/types/Restaurant';
 
+import './RestaurantList.css';
 class RestaurantList extends BaseComponent {
   #restaurantList;
   #restaurantDBService: RestaurantDBService;
