+++ src/index.js
@@ -1,19 +1,10 @@
 import '@/components/Header/Header';
 import '@/components/FilterContainer/FilterContainer';
-import '@/components/SelectBox/SelectBox';
+import '@/components/Basic/SelectBox/SelectBox';
 import '@/components/RestaurantList/RestaurantList';
-import '@/components/RestaurantItem/RestaurantItem';
-import '@/components/BasicModal/BasicModal';
+import '@/components/RestaurantList/RestaurantItem/RestaurantItem';
+import '@/components/Basic/BasicModal/BasicModal';
 import '@/components/NewRestaurantModal/NewRestaurantModal';
-import '@/components/BasicButton/BasicButton';
+import '@/components/Basic/BasicButton/BasicButton';
 
 import '@/css/index.css';
-
-import '@/components/Header/Header.css';
-import '@/components/FilterContainer/FilterContainer.css';
-import '@/components/SelectBox/SelectBox.css';
-import '@/components/RestaurantList/RestaurantList.css';
-import '@/components/RestaurantItem/RestaurantItem.css';
-import '@/components/BasicModal/BasicModal.css';
-import '@/components/NewRestaurantModal/NewRestaurantModal.css';
-import '@/components/BasicButton/BasicButton.css';
